+++ web/js/ui/util/MessageBox.tsx
@@ -2,10 +2,10 @@ import * as React from 'react';
 import Button from 'reactstrap/lib/Button';
 import {RendererAnalytics} from '../../ga/RendererAnalytics';
 import {LeftRightSplit} from '../left_right_split/LeftRightSplit';
-import {Nav} from '../util/Nav';
+import {Nav} from './Nav';
 import {SURVEY_LINK} from '../../../../apps/repository/js/splash/splashes/survey/Survey';
 
-export class BottomBar extends React.Component<IProps, IState> {
+export class MessageBox extends React.Component<IProps, IState> {
 
     constructor(props: any, context: any) {
         super(props, context);
@@ -19,11 +19,23 @@ export class BottomBar extends React.Component<IProps, IState> {
 
         const width = this.props.width || 600;
 
-        return <div style={{
-                        position: 'fixed',
-                        bottom: '25px',
-                        width: '100%'
-                     }}>
+        const position = this.props.position || 'bottom';
+
+        const style: React.CSSProperties = {
+            position: 'fixed',
+            width: '100%'
+        };
+
+        switch (position) {
+            case "bottom":
+                style.bottom = '25px';
+                break;
+            case "top":
+                style.top = '25px';
+                break;
+        }
+
+        return <div style={style}>
 
             <div style={{
                         width:  width + 'px',
@@ -45,6 +57,7 @@ export class BottomBar extends React.Component<IProps, IState> {
 
 interface IProps {
     readonly width?: number;
+    readonly position?: 'bottom' | 'top';
 }
 
 interface IState {